+++ src/xmldiff/format.py
@@ -19,6 +19,7 @@
 this module provides classes to format the native tree2tree output
 """
 
+import logging
 import sys
 try:
     from xml.dom import EMPTY_NAMESPACE as NO_NS
@@ -32,13 +33,16 @@ class AbstractFormatter(object):
     """
     Formatter interface
     """
+    def __init__(self, stream=None):
+        self.edit_s = []
+        self._stream = stream
 
     def init(self, stream=None):
         """ method called before the begining of the tree 2 tree correction """
-        if stream is None:
-            stream = sys.stdout
-        self.edit_s = []
-        self._stream = stream
+        logging.warning("The init() method of Formatters is deprecated. Set the "
+                        "stream with __init__() instead.")
+        if self._stream is None and stream is not None:
+            self._stream = stream
 
     def add_action(self, action):
         """ method called when an action is added to the edit script """
@@ -75,13 +79,17 @@ class InternalPrinter(AbstractFormatter):
         """
         See AbstractFormatter interface
         """
+        stream = self._stream
+        if stream is None:
+            stream = sys.stdout
+
         if len(action) > 2 and isinstance(action[A_N2], list):
-            self._stream.write('[%s, %s,\n' % (action[A_DESC], action[A_N1]))
-            xml_print(action[A_N2], stream=self._stream)
-            self._stream.write("]\n")
+            stream.write('[%s, %s,\n' % (action[A_DESC], action[A_N1]))
+            xml_print(action[A_N2], stream=stream)
+            stream.write("]\n")
         elif len(action) > 2:
-            self._stream.write('[%s, %s, %s]\n' % (action[A_DESC],
+            stream.write('[%s, %s, %s]\n' % (action[A_DESC],
                                                    action[A_N1],
                                                    action[A_N2]))
         else:
-            self._stream.write('[%s, %s]\n' % (action[A_DESC], action[A_N1]))
+            stream.write('[%s, %s]\n' % (action[A_DESC], action[A_N1]))